+++ util/gin_session.go
@@ -7,15 +7,15 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-func SaveAdminIDToSession(ctx *gin.Context, id uint) {
+func SaveAdminIDToSession(ctx *gin.Context, id int64) {
 	session := sessions.Default(ctx)
 	session.Set(g.AdminIdKey, id)
 	session.Save()
 }
 
-func GetAdminIDFormSession(ctx *gin.Context) uint {
+func GetAdminIDFormSession(ctx *gin.Context) int64 {
 	session := sessions.Default(ctx)
-	uid, _ := session.Get(g.AdminIdKey).(uint)
+	uid, _ := session.Get(g.AdminIdKey).(int64)
 	return uid
 }
